key: extract error on read fail
Before this patch, read fail had generated a Go error to return it to a
user, even though read errors (for example, caused by bad input or wrong
password) were placed in error queue. Since goroutines are not binded to
a thread and error queue is per thread, sometimes it had resulted in
valid connections failing with error. This patch fixes the issue.